# models/server.go
/*
 * DBworkBench
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

import (
	"encoding/json"
	"io"
	"io/ioutil"
)

type Server struct {
	Name string `json:"name" bson:"name"`

	Uri string `json:"uri,omitempty" bson:"uri"`
}

type Servers []Server

func (u *Server) Byte() (objBytes []byte, err error) {
	return json.Marshal(u)
}

func (u *Servers) Byte() (objBytes []byte, err error) {
	return json.Marshal(u)
}

func (u *Server) Merge(eu *Server) (err error) {
	return err
}

func (u *Servers) Parse(bodyReader io.Reader) error {
	body, err := ioutil.ReadAll(bodyReader)
	if err != nil {
		return err
	}

	return json.Unmarshal(body, u)
}

func (u *Server) Parse(bodyReader io.Reader) error {
	body, err := ioutil.ReadAll(bodyReader)
	if err != nil {
		return err
	}

	return json.Unmarshal(body, u)
}